+++ src/constants/dataTestIds.enum.ts
@@ -1,6 +1,8 @@
 export enum DataTestIdsEnum {
   addressTableContainer = 'addressTableContainer',
   confirmBtn = 'confirmBtn',
+  confirmData = 'confirmData',
+  confirmScCall = 'confirmScCall',
   dappModal = 'dappModal',
   ledgerConfirmAddress = 'ledgerConfirmAddress',
   ledgerConnectBtn = 'ledgerConnectBtn',